+++ plutarch-api/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
   `punionResolvingCollisionsWithData`, `pvalueOf`, `passertSorted`,
   `padaSymbolData`, `plovelaceValueOf`, `pisAdaOnlyValue`, `padaOnlyValue`, 
   `pnoAdaValue` to `Plutarch.Api.Value`
+* `PTryFrom` instance for `PData (PAsData (PValue Unsorted NonZero))`
 
 ## 2.1.0 -- 22-01-2024